+++ src/ContentCopy/ContentCopy.ps1
@@ -37,7 +37,6 @@ try {
         Install-Module EpiCloud -Scope CurrentUser -Force
     } else {
         Write-Host "EpiCloud installed."
-        Get-Module -Name EpiCloud -ListAvailable
     }
 
     Write-DxpHostVersion